src/models/ML-training/requirements.txt
pandas
psycopg2-binary
scikit-learn
joblib